backend/pyproject.toml
[project]
name = "bulk-data-processor-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.2",
    "cryptography>=42.0.0",
    "fastapi>=0.128.0",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.7.0",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.0.0",
    "email-validator>=2.0.0",
    "pymysql>=1.1.0",
    "pymongo>=4.6.0",
    "boto3>=1.34.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "requests-mock>=1.12.1",
]
